requirements.txt
absl-py==0.13.0
alembic==1.4.2
asgiref==3.4.1
astroid==2.4.2
astunparse==1.6.3
attrs==21.2.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
clang==5.0
click==7.1.2
colorama==0.4.3
cycler==0.10.0
Cython==0.29.24
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Blogging==1.2.2
Flask-Caching==1.9.0
Flask-FileUpload==0.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
gunicorn==20.0.4
h5py==3.1.0
idna==3.2
imageio==2.9.0
imantics==0.1.12
imgaug==0.4.0
imutils==0.5.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
labelme2coco==0.1.2
lazy-object-proxy==1.4.3
lxml==4.6.3
Mako==1.1.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.3
mccabe==0.6.1
networkx==2.6.2
numpy==1.21.2
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
opt-einsum==3.3.0
Pillow==8.3.1
pixellib==0.6.6
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.18.0
pytesseract==0.3.8
python-dateutil==2.8.2
python-editor==1.0.4
python-slugify==4.0.1
pytz==2021.1
PyWavelets==1.1.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.2
scipy==1.7.1
Shapely==1.7.1
shortuuid==1.0.1
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
tesseract==0.1.3
text-unidecode==1.3
tifffile==2021.8.8
toml==0.10.1
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==2.3.3
xmljson==0.2.1
